requirements.txt
transformers==4.37.2
flash-attn==2.4.2
trl==0.7.7
accelerate==0.33
peft
sentencepiece
openai==0.28
tiktoken
colorama
einops
wandb

protobuf
aiohttp
fastapi
httpx
markdown2[all]
nh3
prompt_toolkit>=3.0.0
pydantic<2,>=1
rich>=10.0.0
shortuuid
uvicorn
backoff

ai2thor==2.1.0
opencv-python
textworld[pddl]>=1.6.1
alfworld==0.3.5
textworld==1.6.1
gdown
cleantext==1.1.4
selenium==4.2.0
gym==0.23.1
rank-bm25==0.2.2
werkzeug==2.3.8
thefuzz==0.20.0
scikit_learn==1.1.1
requests_mock